[wallet/mobile] task: refactor AddressBook and ContactItem components
