+++ src/screens/AddressBook.js
@@ -1,9 +1,8 @@
-import React, { Component } from 'react';
-import { FlatList, StyleSheet } from 'react-native';
-import { Button, GradientBackground, ListContainer, ListItem, Section, Tabs, TitleBar } from '@src/components';
+import React, { useState } from 'react';
 import { connect } from 'react-redux';
+import { FlatList, StyleSheet, TouchableOpacity } from 'react-native';
+import { Button, ContactItem, GradientBackground, ListContainer, ListItem, Section, Tabs, TitleBar } from '@src/components';
 import { Router } from '@src/Router';
-import Contact from '@src/components/organisms/Contact';
 import store from '@src/store';
 import translate from '@src/locales/i18n';
 
@@ -12,64 +11,68 @@ const styles = StyleSheet.create({
         marginBottom: 26,
     },
 });
-class AddressBookPage extends Component {
-    state = {
-        listType: 'whitelist',
-    };
 
-    submit = () => {
+export function AddressBookPage(props) {
+    const { addressBook, componentId } = props;
+    const [listType, setListType] = useState('whitelist');
+    const tabs = [
+        {
+            value: 'whitelist',
+            label: translate('addressBook.whitelist'),
+        },
+        {
+            value: 'blacklist',
+            label: translate('addressBook.blacklist'),
+        },
+    ];
+    const contactList = addressBook
+        .getAllContacts()
+        .filter(contact => (listType === 'whitelist' ? !contact.isBlackListed : contact.isBlackListed));
+
+    const goBack = () => Router.goBack(componentId);
+    const openContact = contact =>
+        store
+            .dispatchAction({
+                type: 'addressBook/selectContact',
+                payload: contact,
+            })
+            .then(() => Router.goToContactProfile({}, componentId));
+    const addContact = () =>
         store
             .dispatchAction({
                 type: 'addressBook/selectContact',
                 payload: null,
             })
-            .then(() => Router.goToAddContact({}, this.props.componentId));
-    };
+            .then(() => Router.goToAddContact({}, componentId));
 
-    render() {
-        const { addressBook, componentId } = this.props;
-        const { listType } = this.state;
-        const tabs = [
-            {
-                value: 'whitelist',
-                label: translate('addressBook.whitelist'),
-            },
-            {
-                value: 'blacklist',
-                label: translate('addressBook.blacklist'),
-            },
-        ];
-        const contactList = addressBook
-            .getAllContacts()
-            .filter(contact => (listType === 'whitelist' ? !contact.isBlackListed : contact.isBlackListed));
-
-        return (
-            <GradientBackground
-                name="mesh"
-                theme="light"
-                titleBar={<TitleBar theme="light" title={translate('addressBook.title')} onBack={() => Router.goBack(componentId)} />}
-            >
-                <Tabs style={styles.tabs} value={listType} list={tabs} onChange={listType => this.setState({ listType })} />
-                <ListContainer isScrollable={false}>
-                    <FlatList
-                        data={contactList}
-                        renderItem={({ item, index }) => (
+    return (
+        <GradientBackground
+            name="mesh"
+            theme="light"
+            titleBar={<TitleBar theme="light" title={translate('addressBook.title')} onBack={goBack} />}
+        >
+            <Tabs style={styles.tabs} value={listType} list={tabs} onChange={setListType} />
+            <ListContainer isScrollable={false}>
+                <FlatList
+                    data={contactList}
+                    renderItem={({ item, index }) => (
+                        <TouchableOpacity onPress={() => openContact(item)}>
                             <ListItem>
-                                <Contact data={item} componentId={componentId} id={item.id} key={listType + index} />
+                                <ContactItem data={item} key={listType + index} />
                             </ListItem>
-                        )}
-                    />
-                </ListContainer>
-                <Section type="form-bottom">
-                    <Section type="list">
-                        <Section type="form-item">
-                            <Button text={translate('addressBook.addContact')} theme="light" onPress={() => this.submit()} />
-                        </Section>
+                        </TouchableOpacity>
+                    )}
+                />
+            </ListContainer>
+            <Section type="form-bottom">
+                <Section type="list">
+                    <Section type="form-item">
+                        <Button text={translate('addressBook.addContact')} theme="light" onPress={addContact} />
                     </Section>
                 </Section>
-            </GradientBackground>
-        );
-    }
+            </Section>
+        </GradientBackground>
+    );
 }
 
 export default connect(state => ({
